requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.12.1
lxml==5.3.0
pypdf==5.0.1
python-docx==1.1.2
llama-cpp-python==0.2.90
numpy==2.1.1
fastembed==0.3.6
pydantic==2.9.2
